+++ application/src/main/java/com/ericsson/bss/cassandra/ecchronos/application/config/connection/AgentConnectionConfig.java
@@ -105,6 +105,7 @@ public final class AgentConnectionConfig
         return myLocalDatacenter;
     }
 
+
     /**
      * Gets the DataCenterAwarePolicy used for load-balancing policy.
      *